+++ @geogebra/solver-sdk/src/graspable-math/create-path-map.ts
@@ -208,6 +208,7 @@ function annotate(
     case 'AddEquations':
     case 'SubtractEquations':
     case 'EquationUnion':
+    case 'ExpressionWithConstraint':
     case 'StatementWithConstraint':
     case 'Solution':
     case 'Identity':