+++ engine/src/main/kotlin/engine/expressions/ExpressionWithConstraint.kt
@@ -1,0 +1,16 @@
+package engine.expressions
+
+import engine.operators.ExpressionWithConstraintOperator
+
+class ExpressionWithConstraint(
+    expression: Expression,
+    constraint: Expression,
+    meta: NodeMeta = BasicMeta(),
+) : Expression(
+    operator = ExpressionWithConstraintOperator,
+    operands = listOf(expression, constraint),
+    meta = meta,
+) {
+    val expression get() = firstChild
+    val constraint get() = secondChild
+}