+++ engine/src/test/kotlin/engine/parser/ParserTest.kt
@@ -3,6 +3,7 @@ package engine.parser
 import engine.expressions.Constants
 import engine.expressions.Decorator
 import engine.expressions.Expression
+import engine.expressions.ExpressionWithConstraint
 import engine.expressions.FiniteSet
 import engine.expressions.Inequation
 import engine.expressions.Product
@@ -18,6 +19,7 @@ import engine.expressions.derivativeOf
 import engine.expressions.equationOf
 import engine.expressions.explicitProductOf
 import engine.expressions.expressionOf
+import engine.expressions.finiteSetOf
 import engine.expressions.fractionOf
 import engine.expressions.greaterThanEqualOf
 import engine.expressions.greaterThanOf
@@ -46,7 +48,6 @@ import engine.expressions.rawRootOf
 import engine.expressions.setSolutionOf
 import engine.expressions.setUnionOf
 import engine.expressions.sinOf
-import engine.expressions.solutionSetOf
 import engine.expressions.squareBracketOf
 import engine.expressions.squareRootOf
 import engine.expressions.statementUnionOf
@@ -398,6 +399,14 @@ class ParserTest {
     }
 
     @Test
+    fun testExpressionWithConstraint() {
+        parsesTo(
+            "x + 1 GIVEN x > 0",
+            ExpressionWithConstraint(sumOf(xp("x"), xp(1)), greaterThanOf(xp("x"), xp(0))),
+        )
+    }
+
+    @Test
     fun testSets() {
         parsesTo("/reals/", Constants.Reals)
         parsesTo(
@@ -425,19 +434,19 @@ class ParserTest {
         )
         parsesTo(
             "SetSolution[x: {1}]",
-            setSolutionOf(variableListOf("x"), solutionSetOf(xp(1))),
+            setSolutionOf(variableListOf("x"), finiteSetOf(xp(1))),
         )
         parsesTo(
             "SetSolution[x, y : {(1, 2)}]",
-            setSolutionOf(variableListOf("x", "y"), solutionSetOf(tupleOf(xp(1), xp(2)))),
+            setSolutionOf(variableListOf("x", "y"), finiteSetOf(tupleOf(xp(1), xp(2)))),
         )
         parsesTo(
             "SetSolution[x, y : {1} * /reals/]",
-            setSolutionOf(variableListOf("x", "y"), cartesianProductOf(solutionSetOf(xp(1)), Constants.Reals)),
+            setSolutionOf(variableListOf("x", "y"), cartesianProductOf(finiteSetOf(xp(1)), Constants.Reals)),
         )
         parsesTo(
             "SetSolution[x : SetUnion[{1}, (2, 3)]]",
-            setSolutionOf(variableListOf("x"), setUnionOf(solutionSetOf(xp(1)), openIntervalOf(xp(2), xp(3)))),
+            setSolutionOf(variableListOf("x"), setUnionOf(finiteSetOf(xp(1)), openIntervalOf(xp(2), xp(3)))),
         )
     }