+++ methods/src/main/kotlin/methods/algebra/AlgebraExplanation.kt
@@ -11,6 +11,54 @@ enum class AlgebraExplanation : CategorisedMetadataKey {
      */
     SimplifyAlgebraicExpression,
 
+    /**
+     * Compute the domain on which a given algebraic expression is defined (e.g. exclude zero
+     * denominators) then simplify it.
+     */
+    ComputeDomainAndSimplifyAlgebraicExpression,
+
+    /**
+     * Compute the domain on which a given algebraic expression is defined (e.g. exclude zero
+     * denominators).
+     */
+    ComputeDomainOfAlgebraicExpression,
+
+    /**
+     * Combine the domain of the expression (e.g. x > 0) with the simplified expression (e.g. x + 1) to get
+     * an expression with a constraint (e.g. x + 1 given that x > 0).
+     */
+    CombineSimplifiedExpressionWithConstraint,
+
+    /**
+     * The given expression is defined for all real numbers.
+     */
+    ExpressionIsDefinedEverywhere,
+
+    /**
+     * States that a denominator cannot be zero because it would make the fraction undefined.
+     *
+     * %0 - the denominator
+     * %1 - the fraction which would become undefined
+     */
+    DenominatorMustNotBeZero,
+
+    /**
+     * States that a divisor cannot be zero because it would make the division undefined.
+     *
+     * %0 - the divisor
+     * %1 - the division which would become undefined
+     */
+    DivisorMustNotBeZero,
+
+    /**
+     * Collect domain restrictions into one.
+     *
+     * E.g. for x/(x - 1) + x/(x - 2)
+     *      for the first fraction we get x != 1
+     *      for the second fraction we get x != 2
+     *      and collecting the restrictions results in x != 1 and x != 2
+     */
+    CollectDomainRestrictions,
     ;
 
     override val category = "Algebra"