+++ methods/src/main/kotlin/methods/solvable/SolvableUtils.kt
@@ -5,118 +5,71 @@ import engine.expressions.Constants
 import engine.expressions.Contradiction
 import engine.expressions.Expression
 import engine.expressions.ExpressionComparator
-import engine.expressions.FiniteSet
 import engine.expressions.Identity
-import engine.expressions.Interval
 import engine.expressions.RootOrigin
 import engine.expressions.SetExpression
 import engine.expressions.SetSolution
 import engine.expressions.VariableList
 import engine.expressions.negOf
 import engine.expressions.setSolutionOf
-import engine.expressions.solutionSetOf
 import engine.expressions.sumOf
 import engine.sign.Sign
 import methods.constantexpressions.ConstantExpressionsPlans
 
 fun computeOverallUnionSolution(solutions: List<Expression>): Expression? {
-    // If one of the equations results in an identity, then the overall solution is also an identity
-    val identity = solutions.firstOrNull { it is Identity }
-    if (identity != null) {
-        return identity
+    val solutionSets = solutions.mapNotNull { solution ->
+        when (solution) {
+            // If one of solutions is an identity, then the overall solution is also an identity
+            is Identity -> return solution
+            // Contradictions are irrelevant in a union
+            is Contradiction -> null
+            // Select SetSolutions
+            is SetSolution -> solution.solutionSet
+            // Otherwise we cannot merge the solutions
+            else -> return null
+        }
     }
 
-    val (singleSolutions, intervals) = separateSingleSolutionsAndIntervals(solutions) ?: return null
-
-    return if (intervals.size > 1) {
-        val unionSet = intervals
-            .subList(1, intervals.size)
-            .fold(intervals[0] as SetExpression) { acc, interval ->
-                acc.union(interval, expressionComparator) as SetExpression
-            }
-
-        setSolutionOf(solutions[0].firstChild as VariableList, unionSet)
-    } else if (intervals.size == 1) {
-        // try to merge the single solutions into the interval
-        mergeSolutionsIntoInterval(singleSolutions, intervals[0])?.let {
-            setSolutionOf(
-                solutions[0].firstChild as VariableList,
-                it,
-            )
-        }
-    } else {
-        setSolutionOf(
-            solutions[0].firstChild as VariableList,
-            solutionSetOf(singleSolutions.sortedBy { it.doubleValue }),
-        )
+    return computeUnionOfSets(solutionSets)?.let {
+        setSolutionOf(solutions[0].firstChild as VariableList, it)
     }
 }
 
 fun computeOverallIntersectionSolution(solutions: List<Expression>): Expression? {
-    // If one of the equations results in an identity, then the overall solution is also an identity
-    val identity = solutions.firstOrNull { it is Identity }
-    if (identity != null) {
-        return identity
-    }
-
-    val (singleSolutions, intervals) = separateSingleSolutionsAndIntervals(solutions) ?: return null
-
-    return if (intervals.size > 1) {
-        val intersectionSet = intervals
-            .subList(1, intervals.size)
-            .fold(intervals[0] as SetExpression) { acc, interval ->
-                acc.intersect(interval, expressionComparator) as SetExpression
-            }
-
-        setSolutionOf(solutions[0].firstChild as VariableList, intersectionSet)
-    } else if (intervals.size == 1) {
-        // try to merge the single solutions into the interval
-        mergeSolutionsIntoInterval(singleSolutions, intervals[0])?.let {
-            setSolutionOf(
-                solutions[0].firstChild as VariableList,
-                it,
-            )
-        }
-    } else {
-        setSolutionOf(
-            solutions[0].firstChild as VariableList,
-            solutionSetOf(singleSolutions.sortedBy { it.doubleValue }),
-        )
-    }
-}
-
-fun mergeSolutionsIntoInterval(singleSolutions: Set<Expression>, interval: Interval): Interval? {
-    var currentInterval = interval
-
-    for (solution in singleSolutions) {
-        if (currentInterval.leftBound == solution) {
-            currentInterval = currentInterval.leftClosed()
-        } else if (currentInterval.rightBound == solution) {
-            currentInterval = currentInterval.rightClosed()
-        } else if (currentInterval.contains(solution, expressionComparator) != true) {
-            return null
-        }
-    }
-
-    return currentInterval
-}
-
-private fun separateSingleSolutionsAndIntervals(solutions: List<Expression>): Pair<Set<Expression>, List<Interval>>? {
-    val singleSolutions = mutableSetOf<Expression>()
-    val intervals = mutableListOf<Interval>()
-
-    for (solution in solutions) {
+    val solutionSets = solutions.mapNotNull { solution ->
         when (solution) {
-            is SetSolution -> when (val solutionSet = solution.solutionSet) {
-                is FiniteSet -> singleSolutions.addAll(solutionSet.elements)
-                is Interval -> intervals.add(solutionSet)
-                else -> return null
-            }
-            !is Contradiction -> return null
+            // If one of solutions is a contradiction, then the overall solution is also a contradiction
+            is Contradiction -> return solution
+            // Identities are irrelevant in an intersection
+            is Identity -> null
+            // Select SetSolutions
+            is SetSolution -> solution.solutionSet
+            // Otherwise we cannot merge the solutions
+            else -> return null
         }
     }
 
-    return Pair(singleSolutions, intervals)
+    return computeIntersectionOfSets(solutionSets)?.let {
+        setSolutionOf(solutions[0].firstChild as VariableList, it)
+    }
+}
+
+fun computeUnionOfSets(sets: List<SetExpression>): Expression? {
+    return when (sets.size) {
+        0 -> Constants.EmptySet
+        else -> sets.reduce { acc, set ->
+            acc.union(set, expressionComparator) ?: return null
+        }
+    }
+}
+
+fun computeIntersectionOfSets(sets: List<SetExpression>): Expression? {
+    return when (sets.size) {
+        0 -> Constants.Reals
+        else -> sets.reduce { acc, set ->
+            acc.intersect(set, expressionComparator) ?: return null
+        }
+    }
 }
 
 val expressionComparator = ExpressionComparator { e1: Expression, e2: Expression ->
@@ -128,8 +81,7 @@ val expressionComparator = ExpressionComparator { e1: Expression, e2: Expression
         else -> {
             val diff = sumOf(e1, negOf(e2)).withOrigin(RootOrigin())
             val result = ConstantExpressionsPlans.SimplifyConstantExpression.tryExecute(emptyContext, diff)
-                ?: return@ExpressionComparator Sign.UNKNOWN
-            val simplifiedDiff = result.toExpr
+            val simplifiedDiff = result?.toExpr ?: diff
             val signOfDiff = simplifiedDiff.signOf()
             if (signOfDiff != Sign.UNKNOWN) {
                 signOfDiff